ref: client/server split
